Updated ViewModel

+++ PhotoGallery/Controllers/PhotosController.cs
@@ -14,12 +14,12 @@ namespace PhotoGallery.Controllers
     public class PhotosController : Controller
     {
         private ApplicationDbContext db = new ApplicationDbContext();
-       
+
 
         // GET: Photos/Gallery
         public ActionResult Gallery()
         {
-            ViewBag.Message = "Gallery";    
+            ViewBag.Message = "Gallery";
             using (db)
             {
                 var photos = db.Photos
@@ -27,7 +27,7 @@ namespace PhotoGallery.Controllers
                     .ToList();
 
                 return View(photos);
-            }  
+            }
         }
         public ActionResult newGallery()
         {
@@ -46,28 +46,35 @@ namespace PhotoGallery.Controllers
         public ActionResult Details(int? id)
         {
             ViewBag.Message = "Other";
-          
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Photo photo = db.Photos.Find(id);
-            if (photo == null)
+
+            using (db)
             {
-                return HttpNotFound();
+                var photo = db.Photos
+                    .Where(p => p.Id == id)
+                    .Include(p => p.Author)
+                    .First();
+
+                if (photo == null)
+                {
+                    return HttpNotFound();
+                }
+
+                return View(photo);
             }
-            
-            return View(photo);
         }
 
         // GET: Photos/Create
         [Authorize]
         public ActionResult Create()
         {
-             Photo photo = new Photo();
             ViewBag.Message = "PostImage";
-          
-            return View(photo);
+
+            return View();
         }
 
         // POST: Photos/Create
@@ -76,21 +83,24 @@ namespace PhotoGallery.Controllers
         [HttpPost]
         [Authorize]
         [ValidateInput(false)]
-        public ActionResult Create([Bind(Include = "Id,Title,DateAdded")] Photo photo,HttpPostedFileBase file)
+        public ActionResult Create(Photo photo)
         {
             ViewBag.Message = "PostImage";
-            if (file != null)
+
+            using (db)
             {
-               photo.Image = new byte[file.ContentLength];
-                file.InputStream.Read(photo.Image, 0, file.ContentLength);
+                var authorId = db.Users
+                    .Where(u => u.UserName == this.User.Identity.Name)
+                    .First()
+                    .Id;
+
+                photo.AuthorId = authorId;
+
+                db.Photos.Add(photo);
+                db.SaveChanges();
+
+                return RedirectToAction("MyGallery", "Photos");
             }
-
-            photo.Author = db.Users.FirstOrDefault(u => u.Email == User.Identity.Name);
-
-            db.Photos.Add(photo);
-            db.SaveChanges();
-            
-            return RedirectToAction("MyGallery","Photos");
         }
 
         // GET: Photos/Edit/5
@@ -102,18 +112,26 @@ namespace PhotoGallery.Controllers
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Photo photo = db.Photos.Find(id);
 
-            //if (!IsUserAuthorizedToEdit(photo))
-            //{
-
-            //}
-
-            if (photo == null)
+            using (db)
             {
-                return HttpNotFound();
+                var photo = db.Photos
+                    .Where(p => p.Id == id)
+                    .First();
+
+                if (photo == null)
+                {
+                    return HttpNotFound();
+                }
+
+                var model = new PhotoViewModel();
+                model.Id = photo.Id;
+                model.AuthorId = photo.AuthorId;
+                model.Title = photo.Title;
+                model.Image = photo.Image;
+
+                return View(photo);
             }
-            return View(photo);
         }
 
         // POST: Photos/Edit/5
@@ -122,27 +140,28 @@ namespace PhotoGallery.Controllers
         [HttpPost]
         [Authorize]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit(Photo model)
+        public ActionResult Edit(PhotoViewModel model)
         {
 
             ViewBag.Message = "Other";
-         
 
-                using (var database = new ApplicationDbContext())
-                {
-                    var photo = database.Photos
-                        .FirstOrDefault(p => p.Id == model.Id);
 
-                    photo.Title = model.Title;
-                    DateTime currentDate=DateTime.Now;
-                    photo.DateAdded = currentDate;
+            using (var database = new ApplicationDbContext())
+            {
+                var photo = database.Photos
+                    .FirstOrDefault(p => p.Id == model.Id);
 
-                    database.Entry(photo).State = EntityState.Modified;
-                    database.SaveChanges();
+                photo.AuthorId = model.AuthorId;
+                photo.Title = model.Title;
+                DateTime currentDate = DateTime.Now;
+                photo.DateAdded = currentDate;
+                photo.Image = model.Image;
 
-                    return RedirectToAction("MyGallery", "Photos");
-                }
+                database.Entry(photo).State = EntityState.Modified;
+                database.SaveChanges();
 
+                return RedirectToAction("MyGallery", "Photos");
+            }
         }
 
         // GET: Photos/Delete/5
@@ -154,31 +173,51 @@ namespace PhotoGallery.Controllers
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Photo photo = db.Photos.Find(id);
 
-            //if (!IsUserAuthorizedToEdit(photo))
-            //{
-                //return new HttpStatusCodeResult(HttpStatusCode.Forbidden);
-            //}
-
-            if (photo == null)
+            using (db)
             {
-                return HttpNotFound();
+                var photo = db.Photos
+                    .Where(p => p.Id == id)
+                    .Include(a => a.Author)
+                    .First();
+
+                if (photo == null)
+                {
+                    return HttpNotFound();
+                }
+
+                return View(photo);
             }
-            return View(photo);
         }
 
         // POST: Photos/Delete/5
         [Authorize]
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(int id)
+        public ActionResult DeleteConfirmed(int? id)
         {
-            ViewBag.Message = "Other";
-            Photo photo = db.Photos.Find(id);
-            db.Photos.Remove(photo);
-            db.SaveChanges();
-            return RedirectToAction("MyGallery", "Photos");
+            if(id==null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+
+            using (db)
+            {
+                var photo = db.Photos
+                    .Where(p => p.Id == id)
+                    .Include(a => a.Author)
+                    .First();
+
+                if(photo==null)
+                {
+                    return HttpNotFound();
+                }
+
+                db.Photos.Remove(photo);
+                db.SaveChanges();
+
+                return RedirectToAction("MyGallery", "Photos");
+            }
         }
 
         // GET: Photos/MyGallery
@@ -188,12 +227,12 @@ namespace PhotoGallery.Controllers
             ViewBag.Message = "MyGallery";
             return View(db.Photos.Where(u => u.Author.Email == User.Identity.Name).ToList());
         }
-      
+
         public ActionResult ResultGallery(string userName)
         {
-            
+
             ViewBag.Message = "Home";
-            return View(db.Photos.Where(u => u.Author.Id == userName).ToList()); 
+            return View(db.Photos.Where(u => u.Author.Id == userName).ToList());
         }
         protected override void Dispose(bool disposing)
         {
@@ -206,8 +245,8 @@ namespace PhotoGallery.Controllers
 
         //private bool IsUserAuthorizedToEdit(Photo photo)
         //{
-            //bool isAuthor = photo.IsAuthor(this.User.Identity.Name);
-            //return isAuthor;
+        //bool isAuthor = photo.IsAuthor(this.User.Identity.Name);
+        //return isAuthor;
         //}
     }
 }
